+++ src/pages/common/DatasetWorkspace/plugins/init.ts
@@ -5,6 +5,7 @@ import textPlugin from "./text";
 import edfPlugin from "./edf";
 import niftiPlugin from "./nifti";
 import tsvPlugin from "./tsv";
+import wavPlugin from "./wav";
 
 // Register plugins in order of priority
 export const initializePlugins = () => {
@@ -13,5 +14,6 @@ export const initializePlugins = () => {
   registerPlugin(jsonPlugin);
   registerPlugin(edfPlugin);
   registerPlugin(niftiPlugin);
+  registerPlugin(wavPlugin); // Add WAV plugin before default
   registerPlugin(defaultPlugin);
 };