+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changes
 
 ## February 17, 2025
+- Added WAV file plugin with audio playback and waveform visualization
 - Updated URL query parameter mapping between v1 to v2
 - Adjusted home page layout